Fix issue w/ 'after modifying' the selected item This fix ensures that the 'after modifying' function is called ONLY if the 'on modify' function is not called by the advancement.
The advancement calls the same function but sets a field that determines that the function is called by the advancement.

This change is important as to not result in false alarms that can happen if you try to retrieve the selected item(s) from a UI that cannot be accessed via NBT like crafting tables.